Fix addVolume signature for generated Java, order Add menu
