+++ src/main/kotlin/sc/iview/commands/add/AddCylinder.kt
@@ -26,44 +26,53 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package sc.iview.commands.edit.add
+package sc.iview.commands.add
 
 import org.joml.Vector3f
 import org.scijava.command.Command
 import org.scijava.plugin.Menu
 import org.scijava.plugin.Parameter
 import org.scijava.plugin.Plugin
+import org.scijava.util.ColorRGB
 import sc.iview.SciView
-import sc.iview.commands.MenuWeights
-import sc.iview.commands.MenuWeights.EDIT_ADD_SPHERE
+import sc.iview.commands.MenuWeights.ADD
+import sc.iview.commands.MenuWeights.EDIT_ADD_CYLINDER
 
 /**
- * Command to add a sphere in the scene
+ * Command to add a box to the scene
  *
- * @author Kyle Harrington
+ * @author Jan Tiemann
  */
 @Plugin(
     type = Command::class,
     menuRoot = "SciView",
-    menu = [Menu(label = "Add", weight = MenuWeights.ADD), Menu(
-        label = "Sphere...",
-        weight = EDIT_ADD_SPHERE
+    menu = [Menu(label = "Add", weight = ADD), Menu(
+        label = "Cylinder...",
+        weight = EDIT_ADD_CYLINDER
     )]
 )
-class AddSphere : Command {
+class AddCylinder : Command {
+
     @Parameter
     private lateinit var sciView: SciView
 
+    // FIXME
     //    @Parameter
     //    private String position = "0; 0; 0";
-    @Parameter
-    private val radius = 1.0f
 
-    @Parameter(required = false)
-    private val color = SciView.DEFAULT_COLOR
+    @Parameter
+    private var height = 1.0f
+
+    @Parameter
+    private var radius = 1.0f
+
+    @Parameter
+    private var color: ColorRGB = SciView.DEFAULT_COLOR;
+
     override fun run() {
         //final Vector3 pos = ClearGLVector3.parse( position );
-        val pos = Vector3f(0.0f)
-        sciView.addSphere(pos, radius, color)
+        val pos = Vector3f(0f, 0f, 0f)
+
+        sciView.addCylinder(pos,radius,height,color,20)
     }
 }
