+++ src/main/kotlin/sc/iview/commands/add/AddOrientationCompass.kt
@@ -26,7 +26,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package sc.iview.commands.edit.add
+package sc.iview.commands.add
 
 import graphics.scenery.Icosphere
 import graphics.scenery.Mesh
