+++ src/main/kotlin/sc/iview/commands/add/AddPointLight.kt
@@ -6,13 +6,13 @@
  * %%
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
- *
+ * 
  * 1. Redistributions of source code must retain the above copyright notice,
  *    this list of conditions and the following disclaimer.
  * 2. Redistributions in binary form must reproduce the above copyright notice,
  *    this list of conditions and the following disclaimer in the documentation
  *    and/or other materials provided with the distribution.
- *
+ * 
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
  * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
  * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
@@ -26,47 +26,33 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package sc.iview.commands.edit.add
+package sc.iview.commands.add
 
-import graphics.scenery.proteins.Protein
-import graphics.scenery.proteins.RibbonDiagram
-import org.joml.Vector3f
 import org.scijava.command.Command
-import org.scijava.command.DynamicCommand
 import org.scijava.plugin.Menu
 import org.scijava.plugin.Parameter
 import org.scijava.plugin.Plugin
-import org.scijava.ui.UIService
 import sc.iview.SciView
-import sc.iview.commands.MenuWeights
-import java.io.File
+import sc.iview.commands.MenuWeights.ADD
+import sc.iview.commands.MenuWeights.EDIT_ADD_POINTLIGHT
 
 /**
- * Command to add a box to the scene
+ * Command to add a point light to the scene
  *
  * @author Kyle Harrington
  */
-@Plugin(type = Command::class, menuRoot = "SciView", menu = [Menu(label = "Add", weight = MenuWeights.ADD), Menu(label = "Protein from files ...", weight = MenuWeights.EDIT_ADD_BOX)])
-class AddProteinFromFiles : DynamicCommand() {
-
+@Plugin(
+    type = Command::class,
+    menuRoot = "SciView",
+    menu = [Menu(label = "Add", weight = ADD), Menu(
+        label = "Point Light",
+        weight = EDIT_ADD_POINTLIGHT
+    )]
+)
+class AddPointLight : Command {
     @Parameter
-    private lateinit var sciView: SciView
-
-    @Parameter
-    private lateinit var ui: UIService
-
-    @Parameter(label = "Protein files", style = "extensions:pdb/mmtf/cif")
-    private lateinit var files: Array<File>
-
-    @Parameter(label = "Scale")
-    private var scale: Float = 0.1f
-
+    private val sciView: SciView? = null
     override fun run() {
-        files.forEach { file ->
-            val ribbon = RibbonDiagram(Protein.fromID(file.absolutePath))
-            ribbon.name = file.name
-            ribbon.spatial().scale = Vector3f(scale)
-            sciView.addNode(ribbon, true)
-        }
+        sciView!!.addPointLight()
     }
 }
