+++ src/main/kotlin/sc/iview/commands/add/AddVolume.kt
@@ -26,7 +26,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package sc.iview.commands.edit.add
+package sc.iview.commands.add
 
 import net.imagej.Dataset
 import net.imagej.axis.CalibratedAxis
